25882590:review fixes

+++ comments/usecase/core.go
@@ -77,5 +77,5 @@ func (core *Core) GetUserId(ctx context.Context, sid string) (uint64, error) {
 		core.lg.Error("get user id error", "err", err.Error())
 		return 0, fmt.Errorf("get user id err: %w", err)
 	}
-	return uint64(response.Value), nil
+	return response.Value, nil
 }
